Move `try_from_physical` into its own module
`Series::try_from_physical` is general-purpose - it restores any logical dtype
from its physical representation - but it landed in `canonicalize_maps.rs`, whose
name no longer described its contents. Give it `series/ops/from_physical.rs`.

Its tests follow it there. Four of them had ended up in modules they had nothing
to do with: three `from_chunk_and_dtype` tests covering enum codes, objects and
Decimals sat in `logical/map.rs`, testing no Map behaviour at all, and the `Date`
feature-gate regression test sat in `canonicalize_maps.rs`.

Pure code motion.
